requirements.txt
asgiref==3.11.1
Django==6.0.4
django-unfold==0.90.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
mssql-django==1.7.1
PyJWT==2.12.1
pyodbc==5.3.0
pytz==2026.1.post1
sqlparse==0.5.5
tzdata==2026.1
waitress
whitenoise
